gi: Relocate the struct typedefs to the files where the struct is implemented
This way they all end up in the same documentation page when using
smart index, which is most probably what the user wants, if it
is not what he want he can still override it.

Also this allows the GI extension to have the notion of objects inside
the page where the object is implemented in the case the class/instance
structures are defined somewhere else.